project/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch>=2.0.0
torchvision>=0.15.0
medmnist==2.2.0
scikit-learn>=1.2.0
xgboost>=1.7.0
joblib
numpy
pandas
matplotlib
tqdm
Pillow
shap
onnx
onnxruntime
python-multipart
